feat: adding more abstraction examples and utils

+++ android/DioBank/src/main/kotlin/one/digitalinnovation/digionebank/model/Account.kt
@@ -1,13 +1,11 @@
 package one.digitalinnovation.digionebank.model
 
-import java.math.BigDecimal
-
-class Account(val agency: String, val number: String, balance: BigDecimal) {
-    fun deposit(value: BigDecimal) {
+class Account(val agency: String, val number: String, balance: Double) {
+    fun deposit(value: Double) {
         // TODO: make implementation
     }
 
-    fun withdraw(value: BigDecimal) {
+    fun withdraw(value: Double) {
         // TODO: make implementation
     }
 }
